Move all ViewModel code into WizardViewModel.kt
